Partial solution via modifying position_replace
Fixed what was likely a bug, where whether the first completion found
was a nick was used to decide whether to use nick completions on other
completions as well. So completing first with something that was not a
nick would effectively disallow nick completions as options.
This was preventing cycling through options that included non-nicks.

As far as I can tell, `base_word_pos` and `position_replace` are
always the same (except during computation). There's never a time that
I could find when they were different. So we can use `base_word_pos`
to reset `position_replace`, which is cool.

Now with only conditionally modifying `position_replace` by prefix
length we run into the bug of when you don't have the prefix, and you
reduce the length of stuff, then it won't all get replaced. For a 1
char prefix, 1 character is left at the end. (The channel `#nick` and
`Nickname` will produce this, if you add `#` to the list of chars.)