Failing prefix-agnostic completion test case
This test is just like the others, except it's completing "@A" instead
of just "A".

Also some cleanup of other tests, and argument order swapping for the
WEE_TEST_STR macro since I figured out how it's supposed to be invoked.
Moved the adding nicks to nicklist to setup() since it's repeated for
each test now. Also moved the declaration of the `char * str` there too.