Find previous completion using an its index
Previously, the previous completion was found by using string comparison
until the current completion matched the saved `word_found`. This was
problematic if `word_found` ever changed. Now the index of the found
completion is saved, and that's used to determine if the last used
completion has been encountered yet or not.

Added a `word_found_index` field to the `t_gui_completion` struct.

The checking for if other completions exist seems suspect, and I think
it could be improved.

Refactorings:
- added `gui_completion_next_index (int index, int direction)` function
  since there were multiple ternary statements doing this work