Cached other_completion value in t_gui_completion
This lets us remove the need to iterate over the entire list every time
`gui_completion_complete` is called. Instead we check if there's another
completion the first time, and then afterwards we already know if there
are other ones to cycle through.

Another benefit of this is that there's no longer any need to check for
`word_found_seen` either, netting us an entire indentation level.
(Sorry, git diff.)

Simpler code is worth the increased size of the struct.